chore(presence): Move /src/utils into the /frontend directory
The edits that occur now are in the same file as the previous working version
of syncosaurus
