+++ frontend/src/utils/react.js
@@ -27,4 +27,27 @@ function useSubscribe(syncosaurus, query, initial) {
   return data;
 }
 
-export { useSubscribe };
+function usePresence(syncosaurus) {
+  const [presence, setPresence] = useState(syncosaurus.presence);
+  useEffect(() => {
+    syncosaurus.subscribePresence(setPresence);
+  }, [syncosaurus]);
+  return presence || {};
+}
+
+function useUpdateMyPresence(syncosaurus) {
+  useEffect(() => {
+    console.log('setting up presence');
+    const listener = window.addEventListener('mousemove', e => {
+      const mousePosition = { x: e.clientX, y: e.clientY };
+      // console.log('mouse position', mousePosition);
+      syncosaurus.updateMyPresence(mousePosition);
+    });
+
+    return () => {
+      console.log('removing window event listener');
+      removeEventListener('mousemove', listener);
+    };
+  }, [syncosaurus]);
+}
+export { useSubscribe, usePresence, useUpdateMyPresence };
